Makefile

+++ Makefile
@@ -2,9 +2,9 @@ CC ?= cc
 CFLAGS ?=
 LDFLAGS ?=
 TMPDIR ?= /tmp
+VC     ?= ./vc
 
 VCFILE := v.c
-TMPVC  := $(TMPDIR)/vc
 TMPTCC := /var/tmp/tcc
 VCREPO := https://github.com/vlang/vc
 TCCREPO := https://github.com/vlang/tccbin
@@ -13,6 +13,7 @@ GITFASTCLONE := git clone --depth 1 --quiet
 
 #### Platform detections and overrides:
 _SYS := $(shell uname 2>/dev/null || echo Unknown)
+_MACHINE := $(shell uname -m |grep i686 2>/dev/null || echo 'noti686_or_Unknown')
 _SYS := $(patsubst MSYS%,MSYS,$(_SYS))
 _SYS := $(patsubst MINGW%,MinGW,$(_SYS))
 
@@ -45,53 +46,81 @@ endif
 
 all: latest_vc latest_tcc
 ifdef WIN32
-	$(CC) $(CFLAGS) -g -std=c99 -municode -w -o v.exe $(TMPVC)/$(VCFILE) $(LDFLAGS)
-	./v.exe self
+	$(CC) $(CFLAGS) -g -std=c99 -municode -w -o v.exe $(VC)/$(VCFILE) $(LDFLAGS)
+ifdef prod
+	./v.exe -prod self
 else
-	$(CC) $(CFLAGS) -g -std=gnu11 -w -o v $(TMPVC)/$(VCFILE) $(LDFLAGS) -lm
+	./v.exe self
+endif
+else
+	$(CC) $(CFLAGS) -g -std=gnu11 -w -o v $(VC)/$(VCFILE) $(LDFLAGS) -lm -lpthread
 ifdef ANDROID
 	chmod 755 v
 endif
+
+ifdef prod
+	./v -prod self
+else
 	./v self
+endif
+
 ifndef ANDROID
 	$(MAKE) modules
 endif
 endif
-ifdef V_ALWAYS_CLEAN_TMP
-	$(MAKE) clean_tmp
-endif
 	@echo "V has been successfully built"
 	@./v -version
 
-clean: clean_tmp
-	git clean -xf
+#clean: clean_tmp
+#git clean -xf
 
-clean_tmp:
+clean:
+ifneq ($(_MACHINE),i686)
 	rm -rf $(TMPTCC)
-	rm -rf $(TMPVC)
+endif
+	rm -rf $(VC)
 
-latest_vc: $(TMPVC)/.git/config
-	cd $(TMPVC) && $(GITCLEANPULL)
+latest_vc: $(VC)/.git/config
+ifndef local
+	cd $(VC) && $(GITCLEANPULL)
+else
+	@echo "Using local vc"
+endif
 
 fresh_vc:
-	rm -rf $(TMPVC)
-	$(GITFASTCLONE) $(VCREPO) $(TMPVC)
+	$(GITFASTCLONE) $(VCREPO) $(VC)
 
 latest_tcc: $(TMPTCC)/.git/config
 ifndef ANDROID
+ifndef MAC 
+ifneq ($(_MACHINE),i686)
+ifndef local
 	cd $(TMPTCC) && $(GITCLEANPULL)
+else
+	@echo "Using local tcc"
+endif    
+endif
+endif
 endif
 
 fresh_tcc:
 ifndef ANDROID
+ifndef MAC
+ifneq ($(_MACHINE),i686)
 	rm -rf $(TMPTCC)
 	$(GITFASTCLONE) $(TCCREPO) $(TMPTCC)
 endif
+endif
+endif
 
 $(TMPTCC)/.git/config:
+ifndef MAC
+ifneq ($(_MACHINE),i686)
 	$(MAKE) fresh_tcc
+endif
+endif
 
-$(TMPVC)/.git/config:
+$(VC)/.git/config:
 	$(MAKE) fresh_vc
 
 selfcompile:
